requirements.txt
amqp==2.5.2
apispec==3.3.0
appdirs==1.4.3
attrs==19.3.0
billiard==3.6.3.0
celery==4.4.2
certifi==2019.11.28
chardet==3.0.4
click==7.1.1
dateutils==0.6.7
ecdsa==0.15
echopy-lib==0.1.60
Flask==1.1.1
flask-apispec==0.8.7
Flask-InfluxDB==0.3.1
flask-marshmallow==0.11.0
grpcio==1.27.2
grpcio-tools==1.27.2
idna==2.9
importlib-metadata==1.5.0
influxdb==5.2.2
iroha==0.0.5.5
itsdangerous==1.1.0
Jinja2==2.11.1
jsonschema==3.2.0
kombu==4.6.8
MarkupSafe==1.1.1
marshmallow==3.5.1
numpy==1.18.2
openapi-spec-validator==0.2.8
prance==0.18.2
protobuf==3.11.3
psutil==5.7.0
pycryptodome==3.9.7
pylibscrypt==1.8.0
pyrsistent==0.15.7
python-dateutil==2.8.1
pytz==2019.3
PyYAML==5.3.1
redis==3.4.1
requests==2.23.0
semver==2.9.1
six==1.14.0
urllib3==1.25.8
vine==1.3.0
webargs==5.5.3
websocket-client==0.57.0
Werkzeug==1.0.0
zipp==3.1.0
